Not hanging on to old status bar references
