+++ src/main/kotlin/io/unthrottled/doki/ui/status/ThemeStatusBarWidget.kt
@@ -1,7 +1,6 @@
 package io.unthrottled.doki.ui.status
 
 import com.intellij.icons.AllIcons
-import com.intellij.ide.ui.LafManagerListener
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.options.ShowSettingsUtil
@@ -10,9 +9,7 @@ import com.intellij.openapi.ui.popup.ListPopup
 import com.intellij.openapi.wm.StatusBar
 import com.intellij.openapi.wm.StatusBarWidget
 import com.intellij.util.Consumer
-import io.unthrottled.doki.config.THEME_CONFIG_TOPIC
 import io.unthrottled.doki.config.ThemeConfig
-import io.unthrottled.doki.config.ThemeConfigListener
 import io.unthrottled.doki.settings.ThemeSettings.Companion.THEME_SETTINGS_DISPLAY_NAME
 import io.unthrottled.doki.themes.ThemeManager
 import java.awt.event.MouseEvent
@@ -24,26 +21,6 @@ class ThemeStatusBarWidget(private val project: Project) :
   companion object {
     private const val ID = "Doki Theme Status Component"
   }
-
-  private val connect = ApplicationManager.getApplication().messageBus.connect()
-
-  init {
-    connect.subscribe(LafManagerListener.TOPIC, LafManagerListener {
-      if (this::myStatusBar.isInitialized) {
-        myStatusBar.updateWidget(ID)
-      }
-    })
-    val self = this
-    connect.subscribe(THEME_CONFIG_TOPIC, object : ThemeConfigListener {
-      override fun themeConfigUpdated(themeConfig: ThemeConfig) {
-        if (self::myStatusBar.isInitialized) {
-          self.myStatusBar.updateWidget(ID)
-        }
-      }
-    })
-  }
-
-  private lateinit var myStatusBar: StatusBar
 
   override fun getTooltipText(): String? = "Current Theme"
 
@@ -58,18 +35,10 @@ class ThemeStatusBarWidget(private val project: Project) :
   override fun getPresentation(): StatusBarWidget.WidgetPresentation = this
 
   override fun install(statusBar: StatusBar) {
-    myStatusBar = statusBar
     statusBar.updateWidget(ID)
   }
 
-  override fun dispose() {
-    if (this::myStatusBar.isInitialized) {
-      try {
-        myStatusBar.dispose()
-      } catch (e: Throwable) {
-      }
-    }
-  }
+  override fun dispose() {}
 
   override fun getIcon(): Icon? =
     ThemeManager.instance.currentTheme
